Changed double to int.

+++ Jukebox/trunk/src/player/DACPHeckler.java
@@ -197,7 +197,7 @@ public class DACPHeckler {
 		printResponse();
     }
     
-	public void setVolume(double newVolume) {
+	public void setVolume(int newVolume) {
 		send(DACPRequestGenerator.changeVolume(newVolume));
 		printResponse();
 	}
